Create mod cached_state.rs
