+++ src/state/mod.rs
@@ -1,3 +1,4 @@
+pub(crate) mod cached_state;
 pub mod state_api;
 pub mod state_api_objects;
 pub(crate) mod state_chache;
